# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OrientTogo API",
        "description": "Documentation officielle de l'API OrientTogo (Laravel 13). \n    Toutes les requêtes doivent être préfixées par `/api/v1`. \n    L'authentification se fait via Laravel Sanctum (Bearer Token).",
        "contact": {
            "name": "Équipe Technique OrientTogo",
            "email": "contact@orienttogo.tg"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        },
        {
            "url": "/api",
            "description": "Serveur Local (Base Path)"
        }
    ],
    "paths": {
...
        "/v1/careers": {
            "get": {
                "tags": [
                    "Career"
                ],
                "summary": "Liste des careers",
                "description": "Retourne une liste paginée avec filtres optionnels",
                "operationId": "getCareers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d’éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par nom",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CareerResource"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Career"
                ],
                "summary": "Créer un(e) Career",
                "operationId": "createCareer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCareerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Career créé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CareerResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/careers/{career}": {
            "get": {
                "tags": [
                    "Career"
                ],
                "summary": "Détails d’un(e) Career",
                "operationId": "getCareer",
                "parameters": [
                    {
                        "name": "career",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Career trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CareerResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Career"
                ],
                "summary": "Mettre à jour un(e) Career",
                "operationId": "updateCareer",
                "parameters": [
                    {
                        "name": "career",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCareerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Career mis(e) à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CareerResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Career"
                ],
                "summary": "Supprimer un(e) Career",
                "operationId": "deleteCareer",
                "parameters": [
                    {
                        "name": "career",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Career supprimé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé(e)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Une erreur est survenue."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Les données fournies sont invalides."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Le champ est obligatoire."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StoreCareerRequest": {
                "required": [
                    "name",
                    "description",
                    "required_skills",
                    "market_demand",
                    "salary_min",
                    "salary_max",
                    "long_description",
                    "is_promising"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Développeur Fullstack"
                    },
                    "description": {
                        "type": "string",
                        "example": "Création d’applications web..."
                    },
                    "required_skills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "market_demand": {
                        "type": "integer",
                        "example": 9
                    },
                    "salary_min": {
                        "type": "integer",
                        "example": 300000
                    },
                    "salary_max": {
                        "type": "integer",
                        "example": 800000
                    },
                    "long_description": {
                        "type": "string",
                        "example": "..."
                    },
                    "is_promising": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "UpdateCareerRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Développeur Fullstack"
                    },
                    "description": {
                        "type": "string",
                        "example": "Création d’applications web..."
                    },
                    "required_skills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "market_demand": {
                        "type": "integer",
                        "example": 9
                    },
                    "salary_min": {
                        "type": "integer",
                        "example": 300000
                    },
                    "salary_max": {
                        "type": "integer",
                        "example": 800000
                    },
                    "long_description": {
                        "type": "string",
                        "example": "..."
                    },
                    "is_promising": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "CareerResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Développeur Fullstack"
                    },
                    "description": {
                        "type": "string",
                        "example": "Création d’applications web..."
                    },
                    "required_skills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "market_demand": {
                        "type": "integer",
                        "example": 9
                    },
                    "salary_min": {
                        "type": "integer",
                        "example": 300000
                    },
                    "salary_max": {
                        "type": "integer",
                        "example": 800000
                    },
                    "long_description": {
                        "type": "string",
                        "example": "..."
                    },
                    "is_promising": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Utilisez le token retourné lors de la connexion (login ou verify OTP).",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}